+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## v0.4.9
 ### Script updates
 * AutoFarm: Mastery priority now plants seeds with the highest mastery (excluding those with maxed mastery) to maximize mastery and pool xp.
+* Many scripts and utilities fixed by Visua, restoring functionality.
+### UI Changes
+* Menu updates: credits in the SEMI info section
 
 ## v0.4.8
 ## Hot(mess)fix: Melvor has updated to Alpha v0.17! Big changes, big breakage.